+++ castle/apis/request.py
@@ -1,22 +1,22 @@
 import json
 from castle.configuration import configuration
-from castle.session import SessionSharer
+from castle.apis.session import ApisSession
 
 
-class Request(object):
+class ApisRequest(object):
     def __init__(self, headers=None):
         self.headers = headers or dict()
-        self.base_url = Request.build_base_url()
-        self.sharer = SessionSharer()
+        self.base_url = ApisRequest.build_base_url()
+        self.session = ApisSession()
 
     def build_query(self, method, path, params):
-        return self.sharer.session.request(
+        return self.session.get().request(
             method,
             self.build_url(path),
             auth=('', configuration.api_secret),
             timeout=configuration.request_timeout / 1000.0,
             headers=self.headers,
-            verify=Request.verify(),
+            verify=ApisRequest.verify(),
             data=None if params is None else json.dumps(params)
         )