+++ castle/test/api_test.py
@@ -4,7 +4,9 @@ import responses
 from castle.test import unittest
 from castle.api import Api
 from castle.command import Command
-from castle.request import Request
+from castle.apis.request import ApisRequest
+from castle.configuration import configuration
+from castle.exceptions import ConfigurationError
 
 
 def command():
@@ -16,8 +18,14 @@ def response_text():
 
 
 class ApiTestCase(unittest.TestCase):
+    def setUp(self):
+        configuration.api_secret = 'test'
+
+    def tearDown(self):
+        configuration.api_secret = None
+
     def test_init(self):
-        self.assertIsInstance(Api().req, Request)
+        self.assertIsInstance(Api().req, ApisRequest)
 
     @responses.activate
     def test_request(self):
@@ -38,3 +46,9 @@ class ApiTestCase(unittest.TestCase):
             status=200
         )
         self.assertEqual(Api().call(command()), response_text())
+
+    @responses.activate
+    def test_no_api_secret(self):
+        configuration.api_secret = ''
+        with self.assertRaises(ConfigurationError):
+            Api().call(command())